Improve dataset downloading
