+++ configs/matryoshka/bge-semantic-search-config.yaml
@@ -6,7 +6,6 @@ datasets:
   - nfcorpus
   - scidocs
   - webis-touche2020
-datasets_path: datasets
 batch_size: 16
 onnx_filename: model-fp16.onnx
 model_path: model/bge-small-matryoshka
